+++ app/controllers/languages_controller.rb
@@ -15,7 +15,7 @@ class LanguagesController < ApplicationController
   end
 
   def rate_show
-    @apps = @apps.sort_by { |app| app.score.to_i }.reverse
+    @apps = @apps.sort_by { |app| app.score.to_f }.reverse
     @apps = Kaminari.paginate_array(@apps).page(params[:page])
   end
 end
